// twoByTwo/app/src/types/sudokuResponse.ts
export interface SudokuResponse {
  problem?: [[number, number, number, number], [number, number, number, number], [number, number, number, number], [number, number, number, number]];
  solution?: [[number, number, number, number], [number, number, number, number], [number, number, number, number], [number, number, number, number]];
}